+++ Sources/FrameUp/FULayout/AnyFULayout.swift
@@ -32,54 +32,54 @@ public struct AnyFULayout: FULayout {
         contentOffsets(sizes)
     }
     
-//    public static func == (lhs: AnyFULayout, rhs: AnyFULayout) -> Bool {
-//        lhs.fuLayoutName == rhs.fuLayoutName
-//        && lhs.id == rhs.id
-//        && lhs.alignment == rhs.alignment
-//        && lhs.itemAlignment == rhs.itemAlignment
-//        && lhs.fixedSize == rhs.fixedSize
-//        && lhs.maxItemWidth == rhs.maxItemWidth
-//        && lhs.maxItemHeight == rhs.maxItemHeight
-//    }
+    public static func == (lhs: AnyFULayout, rhs: AnyFULayout) -> Bool {
+        lhs.fuLayoutName == rhs.fuLayoutName
+        && lhs.id == rhs.id
+        && lhs.alignment == rhs.alignment
+        && lhs.itemAlignment == rhs.itemAlignment
+        && lhs.fixedSize == rhs.fixedSize
+        && lhs.maxItemWidth == rhs.maxItemWidth
+        && lhs.maxItemHeight == rhs.maxItemHeight
+    }
 }
 
 extension AnyFULayout {
-//    public struct ForEach<Data: RandomAccessCollection, Content: View>: View where Data.Element: Identifiable, Data.Index == Int {
-//        let data: Data
-//        let layout: AnyFULayout
-//        let content: (Data.Element) -> Content
-//        
-//        public init<L: FULayout>(
-//            _ data: Data,
-//            layout: L,
-//            content: @escaping (Data.Element) -> Content
-//        ) {
-//            self.data = data
-//            self.layout = AnyFULayout(layout)
-//            self.content = content
-//        }
-//        
-//        public var body: some View {
-//            layout.forEach(data, content: content)
-//        }
-//    }
-//    
-//    public struct _View<Content: View>: View {
-//        @ViewBuilder
-//        let layout: AnyFULayout
-//        let content: Content
-//        
-//        public init<L: FULayout>(
-//            _ data: Data,
-//            layout: L,
-//            @ViewBuilder content: () -> Content
-//        ) {
-//            self.layout = AnyFULayout(layout)
-//            self.content = content()
-//        }
-//        
-//        public var body: some View {
-//            layout._view { content }
-//        }
-//    }
+    public struct ForEach<Data: RandomAccessCollection, Content: View>: View where Data.Element: Identifiable, Data.Index == Int {
+        let data: Data
+        let layout: AnyFULayout
+        let content: (Data.Element) -> Content
+        
+        public init<L: FULayout>(
+            _ data: Data,
+            layout: L,
+            content: @escaping (Data.Element) -> Content
+        ) {
+            self.data = data
+            self.layout = AnyFULayout(layout)
+            self.content = content
+        }
+        
+        public var body: some View {
+            layout.forEach(data, content: content)
+        }
+    }
+    
+    public struct _View<Content: View>: View {
+        @ViewBuilder
+        let layout: AnyFULayout
+        let content: Content
+        
+        public init<L: FULayout>(
+            _ data: Data,
+            layout: L,
+            @ViewBuilder content: () -> Content
+        ) {
+            self.layout = AnyFULayout(layout)
+            self.content = content()
+        }
+        
+        public var body: some View {
+            layout._view { content }
+        }
+    }
 }
